requirements.txt
pillow>=12.0.0
pymupdf>=1.26.6
pytesseract>=0.3.13
streamlit>=1.51.0
